Hashrate Registers Part 3 (#1301)
* Replace stochastic hashrate with hashrate register

* Calculate heatmap colors

* Adjust hashrate error count tooltop

* Show full precision on axis for hashrate and heap size

* Add interaction settings

Fixes #1222

* Add EMA to hashrate registers

* Review comments

+++ main/screen.c
@@ -77,7 +77,7 @@ static lv_obj_t *wifi_uptime_label;
 
 static lv_obj_t *notification_label;
 
-static double current_hashrate;
+static float current_hashrate;
 static float current_power;
 static uint64_t current_difficulty;
 static float current_chip_temp;
